Move update handling into its own module
